# swagger.yaml
openapi: 3.0.0
info:
  title: Node Tracker API
  version: 1.0.0
  description: REST API for a private BitTorrent tracker with user management, torrents, and IP bans.
  contact:
    name: API Support
    email: support@nodetracker.com
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          description: Unique user ID
        username:
          type: string
          description: Username
        email:
          type: string
          format: email
          description: User email
        role:
          type: string
          enum: [USER, MODERATOR, ADMIN]
          description: User role
        banned:
          type: boolean
          description: Whether the user is banned
        emailVerified:
          type: boolean
          description: Whether the email is verified
        created:
          type: string
          format: date-time
          description: Creation date
        uploaded:
          type: integer
          description: Total bytes uploaded by the user
        downloaded:
          type: integer
          description: Total bytes downloaded by the user
        seedtime:
          type: integer
          description: Total seed time in seconds
        ratio:
          type: number
          format: float
          description: Upload/download ratio (uploaded/downloaded)
...
    UserBan:
      type: object
      properties:
        id:
          type: integer
          description: Unique ban ID
        userId:
          type: integer
          description: Banned user ID
        reason:
          type: string
          description: Ban reason
        bannedBy:
          type: string
          description: Username of admin/moderator who applied the ban
        bannedAt:
          type: string
          format: date-time
          description: Ban date and time
        expiresAt:
          type: string
          format: date-time
          nullable: true
          description: Expiration date (null for permanent ban)
        active:
          type: boolean
          description: Whether the ban is active
        user:
          type: object
          properties:
            id:
              type: integer
            username:
              type: string
            email:
              type: string
              format: email
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
security:
  - bearerAuth: []
...
paths:
...
  /api/users:
    get:
      summary: List all users (admin only)
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Results limit per page
      responses:
        200:
          description: User list with pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        403:
          description: Access denied
  /api/users/{id}:
    get:
      summary: Get specific user information
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: User ID
      responses:
        200:
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        404:
          description: User not found
...
  /api/user-bans/user/{userId}:
    get:
      summary: Get bans for a specific user
      tags: [UserBans]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
          description: User ID
      responses:
        200:
          description: User ban list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserBan'
...